refactor(utils): consolidate object utilities and remove duplicates
- Moved filterObj to shared utils/objectUtils.js
- Added removeUndefinedFields utility for future use
- Updated imports in user and property controllers
- Removed duplicate implementations
