+++ .gitignore
@@ -213,3 +213,4 @@ Thumbs.db
 .rts2_cache_cjs/
 .rts2_cache_es/
 .rts2_cache_umd/
+docker-compose.yml
