--- src/controllers/auth.controller.js
@@ -1,365 +1,0 @@
-import jwt from 'jsonwebtoken';
-import { promisify } from 'util';
-import crypto from 'crypto';
-import User from '../models/User.js';
-import AppError from '../utils/AppError.js';
-import { sendEmail } from '../utils/email.js';
-
-const signToken = (id) => {
-  return jwt.sign({ id }, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_EXPIRES_IN,
-  });
-};
-
-const createSendToken = (user, statusCode, res) => {
-  const token = signToken(user._id);
-  const cookieOptions = {
-    expires: new Date(
-      Date.now() + process.env.JWT_COOKIE_EXPIRES_IN * 24 * 60 * 60 * 1000
-    ),
-    httpOnly: true,
-    secure: process.env.NODE_ENV === 'production',
-  };
-
-  // Remove password from output
-  user.password = undefined;
-
-  res.cookie('jwt', token, cookieOptions);
-
-  res.status(statusCode).json({
-    status: 'success',
-    token,
-    data: {
-      user,
-    },
-  });
-};
-
-const sendVerificationEmail = async (user, req) => {
-  // 1) Generate the random verification token
-  const verificationToken = user.createEmailVerificationToken();
-  await user.save({ validateBeforeSave: false });
-
-  // 2) Send it to user's email
-  const verifyUrl = `${req.protocol}://${req.get('host')}/api/v1/users/verify-email/${verificationToken}`;
-  
-  const message = `Welcome to Dubai Listings! Please verify your email by clicking on the link below:\n\n${verifyUrl}\n\nThis link is valid for 24 hours.`;
-
-  try {
-    await sendEmail({
-      email: user.email,
-      subject: 'Your email verification token (valid for 24 hours)',
-      message,
-    });
-
-    return true;
-  } catch (err) {
-    user.emailVerificationToken = undefined;
-    user.emailVerificationExpires = undefined;
-    await user.save({ validateBeforeSave: false });
-    
-    throw new AppError(
-      'There was an error sending the verification email. Please try again later.',
-      500
-    );
-  }
-};
-
-export const signup = async (req, res, next) => {
-  try {
-    const newUser = await User.create({
-      name: req.body.name,
-      email: req.body.email,
-      password: req.body.password,
-      passwordConfirm: req.body.passwordConfirm,
-      role: req.body.role || 'user',
-      emailVerified: false,
-    });
-
-    // Send verification email
-    await sendVerificationEmail(newUser, req);
-
-    res.status(201).json({
-      status: 'success',
-      message: 'Account created! Please check your email to verify your account.',
-      data: {
-        user: newUser
-      }
-    });
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const login = async (req, res, next) => {
-  try {
-    const { email, password } = req.body;
-
-    // 1) Check if email and password exist
-    if (!email || !password) {
-      return next(new AppError('Please provide email and password!', 400));
-    }
-
-    // 2) Check if user exists && password is correct
-    const user = await User.findOne({ email }).select('+password +emailVerified');
-
-    if (!user || !(await user.correctPassword(password, user.password))) {
-      return next(new AppError('Incorrect email or password', 401));
-    }
-
-    // 3) Check if email is verified
-    if (!user.emailVerified) {
-      // Resend verification email if not verified
-      await sendVerificationEmail(user, req);
-      return next(
-        new AppError('Please verify your email first. A new verification email has been sent.', 401)
-      );
-    }
-
-    // 4) If everything ok, send token to client
-    createSendToken(user, 200, res);
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const verifyEmail = async (req, res, next) => {
-  try {
-    // 1) Get token from params and hash it
-    const hashedToken = crypto
-      .createHash('sha256')
-      .update(req.params.token)
-      .digest('hex');
-
-    // 2) Find user by token and check expiration
-    const user = await User.findOne({
-      emailVerificationToken: hashedToken,
-      emailVerificationExpires: { $gt: Date.now() }
-    });
-
-    if (!user) {
-      return next(new AppError('Token is invalid or has expired', 400));
-    }
-
-    // 3) Update user's email verification status
-    user.emailVerified = true;
-    user.emailVerificationToken = undefined;
-    user.emailVerificationExpires = undefined;
-    await user.save({ validateBeforeSave: false });
-
-    // 4) Log the user in, send JWT
-    createSendToken(user, 200, res);
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const resendVerificationEmail = async (req, res, next) => {
-  try {
-    const { email } = req.body;
-    
-    if (!email) {
-      return next(new AppError('Please provide an email address', 400));
-    }
-
-    const user = await User.findOne({ email });
-    
-    if (!user) {
-      return next(new AppError('No user found with that email address', 404));
-    }
-
-    if (user.emailVerified) {
-      return next(new AppError('Email is already verified', 400));
-    }
-
-    await sendVerificationEmail(user, req);
-
-    res.status(200).json({
-      status: 'success',
-      message: 'Verification email sent! Please check your inbox.'
-    });
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const protect = async (req, res, next) => {
-  try {
-    // 1) Getting token and check if it's there
-    let token;
-    if (
-      req.headers.authorization &&
-      req.headers.authorization.startsWith('Bearer')
-    ) {
-      token = req.headers.authorization.split(' ')[1];
-    } else if (req.cookies.jwt) {
-      token = req.cookies.jwt;
-    }
-
-    if (!token) {
-      return next(
-        new AppError('You are not logged in! Please log in to get access.', 401)
-      );
-    }
-
-    // 2) Verification token
-    const decoded = await promisify(jwt.verify)(token, process.env.JWT_SECRET);
-
-    // 3) Check if user still exists
-    const currentUser = await User.findById(decoded.id);
-    if (!currentUser) {
-      return next(
-        new AppError(
-          'The user belonging to this token no longer exists.',
-          401
-        )
-      );
-    }
-
-    // 4) Check if user changed password after the token was issued
-    if (currentUser.changedPasswordAfter(decoded.iat)) {
-      return next(
-        new AppError('User recently changed password! Please log in again.', 401)
-      );
-    }
-
-    // 5) Check if email is verified
-    if (!currentUser.emailVerified) {
-      return next(
-        new AppError('Please verify your email address to access this resource.', 403)
-      );
-    }
-
-    // GRANT ACCESS TO PROTECTED ROUTE
-    req.user = currentUser;
-    res.locals.user = currentUser;
-    next();
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const restrictTo = (...roles) => {
-  return (req, res, next) => {
-    // roles ['admin', 'agent']. role='user'
-    if (!roles.includes(req.user.role)) {
-      return next(
-        new AppError('You do not have permission to perform this action', 403)
-      );
-    }
-    next();
-  };
-};
-
-export const forgotPassword = async (req, res, next) => {
-  try {
-    // 1) Get user based on POSTed email
-    const user = await User.findOne({ email: req.body.email });
-    if (!user) {
-      return next(new AppError('There is no user with that email address.', 404));
-    }
-
-    // 2) Generate the random reset token
-    const resetToken = user.createPasswordResetToken();
-    await user.save({ validateBeforeSave: false });
-
-    // 3) Send it to user's email
-    const resetURL = `${req.protocol}://${req.get(
-      'host'
-    )}/api/v1/users/resetPassword/${resetToken}`;
-
-    const message = `Forgot your password? Submit a PATCH request with your new password and passwordConfirm to: ${resetURL}.\nIf you didn't forget your password, please ignore this email!`;
-
-    try {
-      await sendEmail({
-        email: user.email,
-        subject: 'Your password reset token (valid for 10 min)',
-        message,
-      });
-
-      res.status(200).json({
-        status: 'success',
-        message: 'Token sent to email!',
-      });
-    } catch (err) {
-      user.passwordResetToken = undefined;
-      user.passwordResetExpires = undefined;
-      await user.save({ validateBeforeSave: false });
-
-      return next(
-        new AppError('There was an error sending the email. Try again later!'),
-        500
-      );
-    }
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const resetPassword = async (req, res, next) => {
-  try {
-    // 1) Get user based on the token
-    const hashedToken = crypto
-      .createHash('sha256')
-      .update(req.params.token)
-      .digest('hex');
-
-    const user = await User.findOne({
-      passwordResetToken: hashedToken,
-      passwordResetExpires: { $gt: Date.now() },
-    });
-
-    // 2) If token has not expired, and there is user, set the new password
-    if (!user) {
-      return next(new AppError('Token is invalid or has expired', 400));
-    }
-
-    user.password = req.body.password;
-    user.passwordConfirm = req.body.passwordConfirm;
-    user.passwordResetToken = undefined;
-    user.passwordResetExpires = undefined;
-    await user.save();
-
-    // 3) Update changedPasswordAt property for the user
-    // 4) Log the user in, send JWT
-    createSendToken(user, 200, res);
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const updatePassword = async (req, res, next) => {
-  try {
-    // 1) Get user from collection
-    const user = await User.findById(req.user.id).select('+password');
-
-    // 2) Check if POSTed current password is correct
-    if (!(await user.correctPassword(req.body.passwordCurrent, user.password))) {
-      return next(new AppError('Your current password is wrong.', 401));
-    }
-
-    // 3) If so, update password
-    user.password = req.body.password;
-    user.passwordConfirm = req.body.passwordConfirm;
-    await user.save();
-
-    // 4) Log user in, send JWT
-    createSendToken(user, 200, res);
-  } catch (err) {
-    next(err);
-  }
-};
-
-export const getMe = (req, res, next) => {
-  req.params.id = req.user.id;
-  next();
-};
-
-export const logout = (req, res) => {
-  res.cookie('jwt', 'loggedout', {
-    expires: new Date(Date.now() + 10 * 1000),
-    httpOnly: true
-  });
-  
-  res.status(200).json({ status: 'success' });
-};
