+++ src/controllers/auth/auth.controller.js
@@ -1,0 +1,88 @@
+import jwt from 'jsonwebtoken';
+import User from '../../models/User.js';
+import AppError from '../../utils/AppError.js';
+
+const createSendToken = (user, statusCode, res) => {
+  const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, {
+    expiresIn: process.env.JWT_EXPIRES_IN,
+  });
+  const cookieOptions = {
+    expires: new Date(Date.now() + process.env.JWT_COOKIE_EXPIRES_IN * 24 * 60 * 60 * 1000),
+    httpOnly: true,
+    secure: process.env.NODE_ENV === 'production',
+  };
+
+  // Remove password from output
+  user.password = undefined;
+
+  res.cookie('jwt', token, cookieOptions);
+
+  res.status(statusCode).json({
+    status: 'success',
+    token,
+    data: {
+      user,
+    },
+  });
+};
+
+export const signup = async (req, res, next) => {
+  try {
+    const newUser = await User.create({
+      name: req.body.name,
+      email: req.body.email,
+      password: req.body.password,
+      passwordConfirm: req.body.passwordConfirm,
+      role: req.body.role || 'user',
+      emailVerified: false,
+    });
+
+    // Send verification email would be handled by a separate email service
+    // await sendVerificationEmail(newUser, req);
+
+    createSendToken(newUser, 201, res);
+  } catch (err) {
+    next(err);
+  }
+};
+
+export const login = async (req, res, next) => {
+  try {
+    const { email, password } = req.body;
+
+    // 1) Check if email and password exist
+    if (!email || !password) {
+      return next(new AppError('Please provide email and password!', 400));
+    }
+
+    // 2) Check if user exists && password is correct
+    const user = await User.findOne({ email }).select('+password +emailVerified');
+
+    if (!user || !(await user.correctPassword(password, user.password))) {
+      return next(new AppError('Incorrect email or password', 401));
+    }
+
+    // 3) Check if email is verified
+    if (!user.emailVerified) {
+      return next(new AppError('Please verify your email before logging in.', 403));
+    }
+
+    // 4) If everything ok, send token to client
+    createSendToken(user, 200, res);
+  } catch (err) {
+    next(err);
+  }
+};
+
+export const logout = (req, res) => {
+  res.cookie('jwt', 'loggedout', {
+    expires: new Date(Date.now() + 10 * 1000),
+    httpOnly: true,
+  });
+  res.status(200).json({ status: 'success' });
+};
+
+export const getMe = (req, res, next) => {
+  req.params.id = req.user.id;
+  next();
+};
