+++ src/controllers/auth/index.js
@@ -1,0 +1,4 @@
+// Export all auth controllers
+export * from './auth.controller.js';
+export * from './password.controller.js';
+export * from './verification.controller.js';
