+++ src/controllers/property/index.js
@@ -1,0 +1,4 @@
+// Export property controllers
+export * from './property.controller.js';
+export * from './property.stats.controller.js';
+// Note: property.utils.js contains internal utilities and is not exported
