+++ src/controllers/review/index.js
@@ -1,0 +1,3 @@
+// Export review controllers and middleware
+export * from './review.controller.js';
+export * from './review.middleware.js';
