+++ src/controllers/user/index.js
@@ -1,0 +1,3 @@
+// Export user controllers and middleware
+export * from './user.controller.js';
+export * from './user.middleware.js';
