+++ src/routes/user.routes.js
@@ -1,51 +1,43 @@
 import express from 'express';
+import { protect, restrictTo } from '../middleware/auth.middleware.js';
+import { updatePassword } from '../controllers/auth';
 import {
   getAllUsers,
   getUser,
   updateMe,
   deleteMe,
-  getMe,
   updateUser,
   deleteUser,
+  setUserIdFromMe,
   uploadUserPhoto,
   resizeUserPhoto,
-} from '../controllers/user.controller.js';
-import {
-  signup,
-  login,
-  logout,
-  protect,
-  restrictTo,
-  forgotPassword,
-  resetPassword,
-  updatePassword,
-  verifyEmail,
-  resendVerificationEmail,
-} from '../controllers/auth.controller.js';
+  checkUserPermission,
+  cleanupUserPhotoOnError,
+} from '../controllers/user';
 
 const router = express.Router();
-
-// Public routes
-router.post('/signup', signup);
-router.post('/login', login);
-router.get('/logout', logout);
-router.post('/forgotPassword', forgotPassword);
-router.patch('/resetPassword/:token', resetPassword);
-router.get('/verify-email/:token', verifyEmail);
-router.post('/resend-verification-email', resendVerificationEmail);
 
 // Protected routes (require authentication)
 router.use(protect);
 
 // User profile routes
-router.get('/me', getMe, getUser);
-router.patch('/updateMyPassword', updatePassword);
-router.patch('/updateMe', uploadUserPhoto, resizeUserPhoto, updateMe);
-router.delete('/deleteMe', deleteMe);
+router.get('/me', setUserIdFromMe, getUser);
+router.patch(
+  '/update-me',
+  uploadUserPhoto,
+  resizeUserPhoto,
+  updateMe,
+  cleanupUserPhotoOnError // Clean up uploaded file if there's an error
+);
+router.delete('/delete-me', deleteMe);
+router.patch('/update-my-password', updatePassword);
 
 // Admin routes
 router.use(restrictTo('admin'));
 router.route('/').get(getAllUsers);
+
+// Apply checkUserPermission middleware to routes that operate on specific users
+router.use('/:id', checkUserPermission);
 router.route('/:id').get(getUser).patch(updateUser).delete(deleteUser);
 
 export default router;
